+++ finality-provider/service/fp_instance.go
@@ -42,8 +42,8 @@ type FinalityProviderInstance struct {
 	criticalErrChan chan<- *CriticalError
 
 	isStarted *atomic.Bool
-
-	wg sync.WaitGroup
+	wg        sync.WaitGroup
+	quit      chan struct{}
 }
 
 // NewFinalityProviderInstance returns a FinalityProviderInstance instance with the given Babylon public key
@@ -159,6 +159,8 @@ func (fp *FinalityProviderInstance) Start(ctx context.Context) error {
 		return fmt.Errorf("failed to start the poller with start height %d: %w", startHeight, err)
 	}
 
+	fp.quit = make(chan struct{})
+
 	fp.wg.Add(2)
 	go fp.finalitySigSubmissionLoop(ctx)
 	go fp.randomnessCommitmentLoop(ctx)
@@ -177,6 +179,7 @@ func (fp *FinalityProviderInstance) Stop() error {
 
 	fp.logger.Info("stopping finality-provider instance", zap.String("pk", fp.GetBtcPkHex()))
 
+	close(fp.quit)
 	fp.wg.Wait()
 
 	fp.logger.Info("the finality-provider instance is successfully stopped", zap.String("pk", fp.GetBtcPkHex()))
@@ -233,7 +236,7 @@ func (fp *FinalityProviderInstance) finalitySigSubmissionLoop(ctx context.Contex
 		select {
 		case <-ticker.C:
 			fp.processAndSubmitSignatures(ctx)
-		case <-ctx.Done():
+		case <-fp.quit:
 			fp.logger.Info("the finality signature submission loop is closing")
 
 			return
@@ -331,7 +334,7 @@ func (fp *FinalityProviderInstance) randomnessCommitmentLoop(ctx context.Context
 		select {
 		case <-ticker.C:
 			fp.processRandomnessCommitment(ctx)
-		case <-ctx.Done():
+		case <-fp.quit:
 			fp.logger.Info("the randomness commitment loop is closing")
 
 			return
@@ -374,9 +377,15 @@ func (fp *FinalityProviderInstance) processRandomnessCommitment(ctx context.Cont
 
 // reportCriticalErr reports a critical error by sending it to the criticalErrChan for further handling.
 func (fp *FinalityProviderInstance) reportCriticalErr(err error) {
-	fp.criticalErrChan <- &CriticalError{
+	select {
+	case fp.criticalErrChan <- &CriticalError{
 		err:     err,
 		fpBtcPk: fp.GetBtcPkBIP340(),
+	}:
+	case <-fp.quit:
+		fp.logger.Debug("skipping error report due to context cancellation", zap.Error(err))
+	default:
+		fp.logger.Error("failed to report critical error (channel full)", zap.Error(err))
 	}
 }