+++ itest/babylon/hmac_e2e_test.go
@@ -1,5 +1,4 @@
 //go:build e2e_babylon
-// +build e2e_babylon
 
 package e2etest_babylon